fix:Dockerをbuildし直さないと修正が反映されなかったのでVolumeマウント設定を追加+微調整
