+++ docker-compose.yml
@@ -2,6 +2,8 @@ version: '3.8'
 services:
   fastapi:
     build: ./fast/.
+    volumes:
+      - ./fast/app:/app
     ports:
       - "8000:8000"
     env_file:
