--- fast/app/alembic/versions/45dea5662e46_initial_migration.py
@@ -1,43 +1,0 @@
-"""initial migration
-
-Revision ID: 45dea5662e46
-Revises: 
-Create Date: 2024-02-19 22:32:33.619442
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = '45dea5662e46'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('ix_users_email', table_name='users')
-    op.drop_index('ix_users_id', table_name='users')
-    op.drop_index('ix_users_name', table_name='users')
-    op.drop_table('users')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('id', mysql.INTEGER(display_width=11), autoincrement=True, nullable=False),
-    sa.Column('name', mysql.VARCHAR(length=255), nullable=True),
-    sa.Column('email', mysql.VARCHAR(length=255), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    mysql_default_charset='latin1',
-    mysql_engine='InnoDB'
-    )
-    op.create_index('ix_users_name', 'users', ['name'], unique=False)
-    op.create_index('ix_users_id', 'users', ['id'], unique=False)
-    op.create_index('ix_users_email', 'users', ['email'], unique=True)
-    # ### end Alembic commands ###
